Move bot presence
Move bot presence to its own file for cleaner integration
